pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "finxnews"
version = "0.1.0"
description = "Daily finance digest from X (Twitter) Recent Search"
requires-python = ">=3.11"
dependencies = [
    "requests>=2.31,<3",
    "pydantic>=2.5,<3",
    "python-dotenv>=1.0,<2",
    "pyyaml>=6.0,<7",
    "openai>=1.12,<2",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4,<9",
    "ruff>=0.2,<1",
    "mypy>=1.8,<2",
    "types-requests",
    "types-PyYAML",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "W", "UP", "B", "SIM"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
